Fix split stack causing black screen in list view
The split stack button from TokenCard was providing an onSplitCompleted
callback that called Navigator.pop(), which was closing the entire list
view instead of just the split sheet. Removed the unnecessary callback
since the split sheet already dismisses itself. The callback is still
used in ExpandedTokenScreen where it's needed to return to the main list.