Fix HiveError: Set properties after item is inserted into box
Moved summoningSick setter call to after tokenProvider.insertItem() because the setter calls save() which requires the object to be in a Hive box. Plain fields like artworkUrl can be set before insert, but properties with setters that call save() must be set after.

This fixes the 'This object is currently not in a box' error during token creation.